requirements.txt
geopy~=2.3.0
matplotlib~=3.7.1
numpy~=1.24.3
pandas~=2.0.1
scipy~=1.10.1
Pillow~=9.5.0
tqdm~=4.65.0
keyboard~=0.13.5